// front/src/types/user.interface.ts
import { IOrder } from "./order.interface"
import { IProduct } from "./product.interface"

export interface IUser {
    id: number
	email: string
	name: string
	avatarPath: string
	phone: string
	address: string
	isAdmin: boolean
}

export interface IUserUpdate {
	email: string
	name?: string
	avatarPath?: string
	phone?: string
	address?: string
}

export interface IFullUser extends IUser {
	favorites: IProduct[]
	orders: IOrder[]
}